+++ R/db.R
@@ -57,11 +57,11 @@ append_db <- function(con, x, tableName) {
   colNames <- names(x =
     DBI::dbGetQuery(conn = con, DBI::sqlInterpolate(conn = con,
       'SELECT * from ?tableName LIMIT 0', tableName = tableName)))
-  x <- x[ , colNames[!colNames %in% 'id']]
+  x <- x[, colNames[!colNames %in% 'id']]
   sf::st_write(obj = x, dsn = con, tableName, append = TRUE)
 }
 
-#' Title
+#' Databse upload module
 #'
 #' @param id
 #'
@@ -122,6 +122,11 @@ dbWriteServer <- function(id, sfObject, tableName, con) {
         sfObject <- sfObject()
         geometryType <- guess_geometry_type(sfObject = sfObject)
         if (geometryType == 'POINT') {
+          # rename incidental transect_id
+          sfObject[['transect_id']] <- add_incidental_ids(
+            transectIds = sfObject[['transect_id']],
+            dataType = sfObject[['data_type']],
+            geometry = sfObject[['geom']])
           # if mapunit is not valid, then it needs review
           needsReview <- !vapply(lapply(sfObject[['mapunit1']],
             validate_membership(members = c(map_unit_veg(), map_unit_non_veg()))
@@ -141,16 +146,341 @@ dbWriteServer <- function(id, sfObject, tableName, con) {
       return(success)
     })
 }
-is_in_db <- function(con, transectIds, tableName) {
-  inDb <- DBI::dbGetQuery(con,
-    sprintf('SELECT DISTINCT transect_id FROM %s', tableName)
+is_in_db <- function(con, transectIds, observers, stagingTable,
+  transectsTable) {
+  if (stagingTable == staging_tables()[['POINT']]) {
+    inDb <- DBI::dbGetQuery(con,
+      sprintf('SELECT DISTINCT t.transect_id, observer
+        FROM (SELECT transect_id, observer FROM %s UNION
+        select transect_id, observer FROM %s) as t',
+        stagingTable, transectsTable)
+    )
+    paste(transectIds, observers, sep = '-') %in%
+      paste(inDb$transect_id, inDb$observer, sep = '-')
+  } else {
+    inDb <- DBI::dbGetQuery(con,
+      sprintf('SELECT DISTINCT t.transect_id
+        FROM (SELECT transect_id FROM %s UNION
+        select transect_id FROM %s) as t',
+        stagingTable, transectsTable)
+    )
+    transectIds %in% inDb$transect_id
+  }
+}
+select_distinct_transect_ids <- function(con, stagingTable, transectsTable) {
+  DBI::dbGetQuery(con,
+    sprintf("SELECT DISTINCT t.transect_id
+      FROM (SELECT transect_id, data_type FROM %s
+      UNION
+      select transect_id, data_type FROM %s ) as t
+      where t.data_type <> 'incidental'", stagingTable,
+      transectsTable)
   )[['transect_id']]
-  transectIds %in% inDb
 }
-filter_in_db <- function(con, tableName, sfObject) {
-  dropTransectIds <- sfObject[['transect_id']][
-    is_in_db(con = con, transectIds = sfObject[['transect_id']],
-      tableName = tableName)]
-  sfObject[!sfObject[['transect_id']] %in% dropTransectIds, ]
+add_incidental_ids <- function(transectIds, dataType, geometry) {
+  if (guess_geometry_type(geometry) == 'POINT') {
+    ifelse(dataType == data_type()['incidental sampling'],
+      apply(X = sf::st_coordinates(geometry), MARGIN = 1,
+        FUN = paste, collapse = ',')
+      , transectIds)
+  } else {
+    transectIds
+  }
 }
+#' PEM database operations
+#'
+#' @param con
+#'
+#' connection to PEM database
+#'
+#' @param transectIds
+#'
+#' transect ids
+#'
+#' @export
+#'
+#' @name dbOperations
+transfer_field_data_points <- function(con, transectIds) {
+  msg <- tryCatch({
+    DBI::dbBegin(conn = con)
+    DBI::dbExecute(con,
+      'CREATE TEMPORARY TABLE validatedFieldPoints (
+      transect_id VARCHAR(255)
+      )
+      ON COMMIT DROP')
+    sql <- DBI::sqlAppendTable(con = con,
+      table = DBI::SQL('validatedFieldPoints'), values =
+        data.frame(transect_id = transectIds), row.names = FALSE)
+    DBI::dbExecute(con, sql)
+    DBI::dbExecute(con,
+      'INSERT INTO transects.field_data_points
+      SELECT fdp.* FROM staging.field_data_points fdp
+      JOIN validatedFieldPoints as vi on fdp.transect_id = vi.transect_id')
+    DBI::dbExecute(con,
+      'DELETE FROM staging.field_data_points as s
+      USING transects.field_data_points t
+      WHERE s.id = t.id')
+    DBI::dbCommit(con)
+  }, error = function(e) {
+    DBI::dbRollback(con)
+    e
+    })
+  msg
+}
+#' @param stagedIds
+#'
+#' integer id of staged tracklogs
+#'
+#' @export
+#'
+#' @name dbOperations
+transfer_field_tracklog <- function(con, stagedIds) {
+  msg <- tryCatch({
+    DBI::dbBegin(conn = con)
+    DBI::dbExecute(con,
+      'CREATE TEMPORARY TABLE validatedTracklog (
+      id INT
+      )
+      ON COMMIT DROP')
+    sql <- DBI::sqlAppendTable(con = con,
+      table = DBI::SQL('validatedTracklog'), values =
+        data.frame(id = stagedIds), row.names = FALSE)
+    DBI::dbExecute(con, sql)
+    DBI::dbExecute(con,
+      'INSERT INTO transects.field_tracklog
+      SELECT ft.* FROM staging.field_tracklog ft
+      JOIN validatedTracklog as vi on ft.id = vi.id')
+    DBI::dbExecute(con,
+      'DELETE FROM staging.field_tracklog as s
+      USING transects.field_tracklog t
+      WHERE s.id = t.id')
+    DBI::dbCommit(con)
+  }, error = function(e) {
+    DBI::dbRollback(con)
+    e
+  })
+  msg
+}
+#' @export
+#'
+#' @rdname dbPhoto
+dbPhotoUi <- function(id) {
+  ns <- shiny::NS(id)
+  shiny::tags$div(
+    shiny::uiOutput(ns('selectTransect')),
+    shiny::uiOutput(ns('selectObserver')),
+    shiny::uiOutput(ns('selectOrder')),
+    shiny::fileInput(ns('photoFile'), label = 'Select Photo', width = '100%',
+      accept = c('jpg', 'png', 'jpeg')),
+    shiny::uiOutput(ns('submitUi')),
+    leaflet::leafletOutput(ns('photoMap'))
+  )
+}
+#' Module to select a point and upload a photo for it
+#'
+#' @param id
+#'
+#' namespace id
+#'
+#' @param con
+#'
+#' connection to postgres database
+#'
+#' @return
+#'
+#' a shiny reactive object row id of photo and a database trigger
+#'
+#' @export
+#'
+#' @name dbPhoto
+#'
+dbPhotoServer <- function(id, con) {
+  shiny::moduleServer(
+    id,
+    function(input, output, session) {
+      ns <- session$ns
+      dbTrigger <- make_reactive_trigger()
+      transectObserver <- shiny::reactive({
+        DBI::dbGetQuery(conn = con,
+          sprintf('SELECT DISTINCT t.transect_id, observer, t.order
+        FROM (SELECT transect_id, observer, "order" FROM %s UNION
+        select transect_id, observer, "order" FROM %s) as t',
+            staging_tables()[['POINT']],
+            transects_tables()[['POINT']]))
+      })
+      output$selectTransect <- shiny::renderUI({
+        shiny::selectizeInput(ns('selectedTransect'), 'Select Transect ID',
+          choices = sort(unique(transectObserver()[['transect_id']])),
+          width = '100%')
+      })
+      output$selectObserver <- shiny::renderUI({
+        observers <- transectObserver()[['observer']][
+          transectObserver()[['transect_id']] == input$selectedTransect]
+        shiny::selectizeInput(ns('selectedObserver'), 'Select Observer',
+          choices = sort(unique(observers)),
+          width = '100%')
+      })
+      output$selectOrder <- shiny::renderUI({
+        orders <- transectObserver()[['order']][
+          transectObserver()[['transect_id']] == input$selectedTransect &
+            transectObserver()[['observer']] == input$selectedObserver]
+        shiny::selectizeInput(ns('selectedOrder'), 'Select Order',
+          choices = sort(unique(orders)),
+          width = '100%')
+      })
+      output$submitUi <- shiny::renderUI({
+        shiny::req(input$selectedTransect, input$selectedObserver,
+          input$selectedOrder, input$photoFile)
+        shiny::actionButton(inputId = ns('submit'), 'Submit',
+          class = 'btn-primary btn-sm mb-3')
+      })
+      mapData <- shiny::reactive({
+        shiny::req(input$selectedTransect, input$selectedObserver)
+        query <- sprintf("SELECT t.*
+        FROM (SELECT * FROM %s UNION
+        select * FROM %s) as t
+            where t.transect_id = \'%s\' and t.observer = \'%s\'",
+          staging_tables()[['POINT']],
+          transects_tables()[['POINT']],
+          input$selectedTransect, input$selectedObserver)
+        sf::st_transform(sf::st_read(dsn = con, query = query), crs = 4326)
+      })
+      output$photoMap <- leaflet::renderLeaflet({
+        shiny::req(mapData(), input$selectedOrder)
+        bb <- sf::st_bbox(mapData())
+        leaflet::leaflet() |>
+          leaflet::addProviderTiles(
+            provider = leaflet::providers$Esri.WorldTopoMap) |>
+          leaflet::fitBounds(lng1 = bb[['xmin']], lat1 = bb[['ymin']],
+            lng2 = bb[['xmax']], lat2 = bb[['ymax']])
 
+      })
+      shiny::observeEvent(input$photoMap_marker_click, {
+        clickId <- stats::setNames(
+          strsplit(input$photoMap_marker_click$id, split = '-')[[1]],
+          c('transect_id', 'observer', 'order'))
+        shiny::updateSelectizeInput(session = session,
+          inputId = 'selectedOrder', selected = clickId[['order']])
+      })
+      shiny::observeEvent(input$selectedOrder, {
+        colors <- ifelse(mapData()[['order']] == input$selectedOrder,
+          zissou()[7], zissou()[1])
+        leaflet::leafletProxy('photoMap') |>
+          leaflet::addCircleMarkers(
+            data = mapData(),
+            radius = 5, color = colors,
+            opacity = .5, fillOpacity = .5, weight = 1,
+            group = 'photoPoints',
+            layerId = ~sprintf('%s-%s-%s', transect_id, observer, order),
+            popup = ~sprintf('%s-%s-%s', transect_id, observer, order))
+      })
+      photoId <- shiny::reactive({
+        shiny::req(input$selectedTransect, input$selectedObserver,
+          input$selectedOrder)
+        get_point_id(con = con, transectId = input$selectedTransect,
+          observer = input$selectedObserver,
+          order = input$selectedOrder)
+      })
+      shiny::observeEvent(input$submit, {
+        shiny::req(input$selectedTransect, input$selectedObserver,
+          input$selectedOrder, input$photoFile)
+        id <- photoId()
+        photo <- blob::as_blob(readBin(input$photoFile$datapath, what = 'raw',
+          n = 1000000))
+        ext <- sub('.+\\.([A-Za-z]+)', '\\1', input$photoFile$datapath)
+        binPhoto <- data.table::data.table(id, photo, ext)
+        writeTransaction <- upsert_photo(con, binPhoto)
+        if (inherits(writeTransaction, 'error')) {
+          shiny::showNotification(ui = writeTransaction$message, type = 'error')
+        } else {
+          shiny::showNotification(ui = 'Photo uploaded.',
+            type = 'default')
+          dbTrigger$trigger()
+        }
+      })
+
+      selectedPoint <- shiny::reactive({
+        list(id = photoId(), success = dbTrigger$depend()
+        )
+      })
+      return(selectedPoint)
+
+    })
+}
+upsert_photo <- function(con, photoDf) {
+  msg <- tryCatch({
+    DBI::dbBegin(conn = con)
+    DBI::dbExecute(con,
+      sprintf('DELETE FROM transects.photos where id = %d', photoDf[['id']]))
+    DBI::dbExecute(conn = con,
+      DBI::sqlAppendTable(con = con, table = DBI::SQL('transects.photos'),
+        values = photoDf, row.names = FALSE))
+    DBI::dbCommit(con)
+  }, error = function(e) {
+    DBI::dbRollback(con)
+    e
+  })
+  msg
+}
+get_point_id <- function(con, transectId, observer, order) {
+  DBI::dbGetQuery(conn = con,
+    sprintf('SELECT id
+        FROM (SELECT "id", transect_id, observer, "order" FROM %s UNION
+        select "id", transect_id, observer, "order" FROM %s) as t
+            where transect_id in (\'%s\') and observer in (\'%s\')
+            and "order" in (%s)',
+      staging_tables()[['POINT']],
+      transects_tables()[['POINT']],
+      transectId, observer, order))[['id']]
+}
+#' @export
+#'
+#' @name ShowPhoto
+dbShowPhotoUi <- function(id) {
+  ns <- shiny::NS(id)
+  shiny::tags$div(shiny::imageOutput(ns('photoUi')),
+    style = 'height: 700px;')
+}
+#' Queries a photo from the database and shows the image
+#'
+#' @param id
+#'
+#' namespace id
+#'
+#' @param con
+#'
+#' connection to postgres
+#'
+#' @param selectedPoint
+#'
+#' reactive list with id of point in database and a database trigger
+#'
+#' @export
+#'
+#' @name ShowPhoto
+#'
+dbShowPhotoServer <- function(id, con, selectedPoint) {
+  shiny::moduleServer(
+    id,
+    function(input, output, session) {
+      output$photoUi <- shiny::renderImage({
+        shiny::req(selectedPoint()[['id']])
+        outfile <- tryCatch({
+          convert_to_image(con, id = selectedPoint()[['id']])
+        }, error = identity)
+        shiny::req(!inherits(outfile, 'error'))
+        list(src = outfile,
+          contentType = 'image/jpg',
+          style = 'display: flex; margin: auto;',
+          height = 700,
+          alt = "This is alternate text")
+      }, deleteFile = TRUE)
+    })
+}
+convert_to_image <- function(con, id) {
+  photo <- DBI::dbGetQuery(con,
+    sprintf('select * from transects.photos where "id" in (%d)', id))
+  temp <- tempfile(fileext = sprintf('.%s', photo[['ext']]))
+  writeBin(photo[['photo']][[1]], con = temp, useBytes = TRUE)
+  temp
+}