+++ R/validate.R
@@ -24,6 +24,10 @@
 #'
 #' TRUE/FALSE for whether missing values are acceptable
 #'
+#' @param progress
+#'
+#' optional shiny progress object
+#'
 #' @param ...
 #'
 #' validation functions for x. See example for writing custom tests.
@@ -58,13 +62,8 @@ validate_data <- function(x, ...) {
     }))
 }
 collect_validations <- function(validations) {
-  # noErrors <- sum(!vapply(validations, FUN = getElement,
-  # FUN.VALUE = logical(1),
-  #   name = 'status'))
   allMessages <- vapply(validations, FUN = getElement, FUN.VALUE = character(1),
     name = 'message')
-  #list(
-    #errors = noErrors,
   c(
     messages = paste0(allMessages[allMessages != ''], collapse = '; ')
   )
@@ -162,8 +161,8 @@ validate_time <- function(optional = FALSE) {
 #' @export
 validate_missing <- function() {
   function(x) {
-    isValid <- is.na(x)
-    success <- any(isValid)
+    isValid <- !is.na(x)
+    success <- all(isValid)
     description <- ifelse(success, '',
       sprintf('Missing records:\n%d records (%.2f%%) missing.',
         sum(!isValid, na.rm = TRUE),
@@ -271,9 +270,8 @@ validate_transect_id <- function() {
       members = transect_location())(transectParts[['part4']])
     isValidSegments <- validate_float()(as.numeric(transectParts[['part2']]))
     isValidTotal <- validate_integer()(as.integer(transectParts[['part3']]))
-    #validPattern <- '[A-Z]{3,4}[a-z]{2}[0-9]*\\_[0-9]{1}\\.[0-9]{1}_[0-9]+\\_c*[A-Za-z]+'
-    validResults <- list(isValidMapUnit, isValidTransectLocation, isValidSegments,
-      isValidTotal)
+    validResults <- list(isValidMapUnit, isValidTransectLocation,
+      isValidSegments, isValidTotal)
     success <- all(vapply(validResults, FUN = getElement,
       FUN.VALUE = logical(1), name = 'status'))
     description <- collect_validations(validResults)
@@ -393,7 +391,10 @@ validate_points_column <- function(dataFrame, colName) {
 #' @rdname validate
 #'
 #' @export
-validate_PEM_data <- function(dataFrame, f) {
+validate_PEM_data <- function(dataFrame, f, progress, n = 1) {
+  if (!missing(progress)) {
+    progress$inc(1 / n)
+  }
   stats::setNames(lapply(names(dataFrame), f, dataFrame = dataFrame),
     names(dataFrame))
 }
@@ -404,10 +405,18 @@ validate_PEM_data <- function(dataFrame, f) {
 #' @rdname validate
 #'
 #' @export
-validate_field_points_data <- function(transectPoints) {
-  lapply(split(transectPoints, transectPoints[['transect_id']]),
-    validate_PEM_data,
-    f = validate_points_column)
+validate_field_points_data <- function(transectPoints, progress) {
+  transectPoints <- split(transectPoints, transectPoints[['transect_id']])
+  if (!missing(progress)) {
+    lapply(transectPoints,
+      validate_PEM_data,
+      f = validate_points_column,
+      progress = progress, n = length(transectPoints))
+  } else {
+    lapply(transectPoints,
+      validate_PEM_data,
+      f = validate_points_column)
+  }
 }
 #' @rdname validate
 #'
@@ -422,7 +431,8 @@ validate_tracklog_column <- function(dataFrame, colName) {
     comments = validate_data(x = dataFrame[[colName]], validate_character(),
       validate_char_length(255)),
     data_type =
-      validate_data(x = dataFrame[[colName]], validate_membership(members = data_type())),
+      validate_data(x = dataFrame[[colName]], validate_membership(
+        members = data_type())),
     geom = validate_data(x = dataFrame[[colName]], #validate_geometry(),
       validate_geometry_type(geometryType = 'LINESTRING')),
     err_column_name(colName)
@@ -435,8 +445,12 @@ validate_tracklog_column <- function(dataFrame, colName) {
 #' @rdname validate
 #'
 #' @export
-validate_field_tracklog_data <- function(transectLines) {
-  lapply(split(transectLines, transectLines[['transect_id']]),
+validate_field_tracklog_data <- function(transectLines, progress, n = 1) {
+  transectLines <- split(transectLines, transectLines[['transect_id']])
+  if (!missing(progress)) {
+    progress$inc(1 / n)
+  }
+  lapply(transectLines,
     validate_PEM_data,
     f = validate_tracklog_column)
 }
@@ -482,7 +496,8 @@ validate_length <- function(n, distinct = TRUE) {
 #' @export
 validate_sample_pairs <- function(transectIds) {
   transectParts <- data.table::tstrsplit(transectIds, '_')
-  transectParts <- expand_list(x = transectParts, n = 4, times = length(transectIds))
+  transectParts <- expand_list(x = transectParts, n = 4,
+    times = length(transectIds))
   lapply(split(transectParts[[4]], paste(transectParts[[1]], transectParts[[2]],
     transectParts[[3]], sep = '_')), validate_data,
     validate_contains(string = 'cLHS', times = 1L, exact = TRUE),
@@ -496,7 +511,19 @@ expand_list <- function(x, n = 4, times) {
     expand_list(x, n = n, times = times)
   }
 }
-#' Shiny Module to validate and report
+#' @export
+#'
+#' @rdname validateTable
+#'
+validateTableUi <- function(id) {
+  ns <- shiny::NS(id)
+  shiny::tags$div(
+    shiny::uiOutput(ns('dbFilterUi')),
+    reactable::reactableOutput(ns('dataValidation'))
+    )
+}
+
+#' Shiny module to validate and report
 #'
 #' @param id
 #'
@@ -504,7 +531,7 @@ expand_list <- function(x, n = 4, times) {
 #'
 #' @param sfObject
 #'
-#' reactive object returning sf data.frame
+#' reactive object returning sf dataframe
 #'
 #' @param success
 #'
@@ -514,58 +541,99 @@ expand_list <- function(x, n = 4, times) {
 #'
 #' connection to postgres database
 #'
-#' @export
-#'
 #' @name validateTable
 #'
-validateTableUi <- function(id) {
-  ns <- shiny::NS(id)
-  shiny::tags$div(
-    reactable::reactableOutput(ns('dataValidation'))
-    )
-}
 #' @export
 #'
-#' @rdname validateTable
 validateTableServer <- function(id, sfObject, success, con) {
   shiny::moduleServer(
     id,
     function(input, output, session) {
-      validationResults <- shiny::reactive({
+      dataInDb <- shiny::reactive({
         shiny::req(sfObject())
         success$depend()
         geometryType <- guess_geometry_type(sfObject())
         stagingTable <- staging_tables()[[geometryType]]
-        sfObject <- tryCatch(
-          filter_in_db(con = con,
-            tableName = stagingTable, sfObject = sfObject()),
+        transectsTable <- transects_tables()[[geometryType]]
+        transectIds <- add_incidental_ids(
+          transectIds = sfObject()[['transect_id']],
+          dataType = sfObject()[['data_type']],
+          geometry = sf::st_geometry(sfObject()))
+        tryCatch(is_in_db(con = con, transectIds = transectIds,
+          observers = sfObject()[['observer']], stagingTable = stagingTable,
+          transectsTable = transectsTable),
           error = identity)
-        if (inherits(x = sfObject, 'error')) {
+      })
+      output$dbFilterUi <- shiny::renderUI({
+        shiny::req(!inherits(dataInDb(), 'error'))
+        geometryType <- guess_geometry_type(sfObject())
+        if (geometryType == 'POINT') {
+          uniqueIds <- unique(as.data.frame(sfObject()[dataInDb() &
+              sfObject()[['data_type']] !=
+              data_type()['incidental sampling'], ])[,
+                c('transect_id', 'observer')])
+          msg <- shiny::tags$div(class = 'mb-2',
+            shiny::tags$h6(
+              'Data already submitted for (transect_id-observer):'),
+            paste(sprintf('%s-%s', uniqueIds[[1]], uniqueIds[[2]]),
+              collapse = '; ')
+          )
+        } else {
+          uniqueIds <- unique(as.data.frame(sfObject()[dataInDb(), ])[,
+            c('transect_id')])
+          msg <- shiny::tags$div(class = 'mb-2',
+            shiny::tags$h6('Data already submitted for (transect_id):'),
+            paste(sprintf('%s', uniqueIds),
+              collapse = '; ')
+          )
+        }
+        if (length(uniqueIds) < 1) {
+          return(NULL)
+        }
+        msg
+      })
+      validationResults <- shiny::reactive({
+        shiny::req(sfObject())
+        success$depend()
+        geometryType <- guess_geometry_type(sfObject())
+        if (inherits(x = dataInDb(), 'error')) {
           shiny::showNotification(
-            ui = 'Network connection error: database is unavailable. Please try again.',
+    ui = 'Network connection error: database is unavailable. Please try again.',
             duration = 5, type = 'error')
           return(NULL)
         }
+        sfObject <- sfObject()[!dataInDb(), ]
+        shiny::req(nrow(sfObject) > 0)
         if (geometryType %in% 'POINT') {
+          progress <- shiny::Progress$new(session)
+          on.exit(progress$close())
+          progress$set(value = .01, message = 'Validation in progress...')
           results <- data.table::rbindlist(
-            lapply(validate_field_points_data(sfObject),
+            lapply(validate_field_points_data(sfObject, progress = progress),
               data.table::as.data.table),
             fill = TRUE, idcol = 'id')
+          hardConstraints <- c('transect_id', 'point_type', 'order',
+            'transition',
+            'observer', 'date_ymd', 'data_type',
+            'geom')
           results[['Valid']] <- vapply(
-            split(results[ , c('transect_id', 'date_ymd', 'data_type', 'geom')],
+            split(results[, hardConstraints],
               results[['id']]), FUN = function(x) {
               all(x == '')
             }, FUN.VALUE = logical(1))
         } else if (geometryType %in% c('LINESTRING', 'MULTILINESTRING')) {
+          progress <- shiny::Progress$new(session)
+          on.exit(progress$close())
+          progress$set(value = .01, message = 'Validation in progress...')
           results <- data.table::rbindlist(
-            lapply(validate_field_tracklog_data(sfObject),
+            lapply(validate_field_tracklog_data(sfObject, progress = progress),
               data.table::as.data.table),
             fill = TRUE, idcol = 'id')
+          hardConstraints <- c('transect_id', 'date_ymd', 'data_type',
+            'geom')
           results[['Valid']] <- vapply(
-            split(results[ , c('order', 'point_type', 'transect_id',
-              #TODO: 'observer', 'transition', 'struc_stage', 'struc_mod',
-              'date_ymd', 'data_type',
-              'geom')], results[['id']]), FUN = function(x) {
+            split(results[, hardConstraints], results[['id']]),
+            FUN = function(x) {
                 all(x == '')
               }, FUN.VALUE = logical(1))
         } else {
@@ -590,7 +658,8 @@ validateTableServer <- function(id, sfObject, success, con) {
             .selection = reactable::colDef(
               sticky = 'left',
               headerClass = 'hide-checkbox',
-              class = ifelse(validationResults()[['Valid']], '', 'hide-checkbox')
+              class = ifelse(validationResults()[['Valid']], '',
+                'hide-checkbox')
             ),
             id = reactable::colDef(
               sticky = 'left',
@@ -616,6 +685,12 @@ validateTableServer <- function(id, sfObject, success, con) {
               }
             }
             ),
+            time_hms = reactable::colDef(cell = function(value) {
+              if (value != '') {
+                shiny::tags$div(class = 'cell-warning', value)
+              }
+            }
+            ),
             Valid =  reactable::colDef(cell = function(value) {
               if (isTRUE(value)) {
                 shiny::tags$div(class = 'cell-success', value)
@@ -632,9 +707,7 @@ validateTableServer <- function(id, sfObject, success, con) {
               if (value != '') {
                 shiny::tags$div(class = 'cell-error', value)
               }
-              #list(background = ifelse(value == '', blue, red))
             }),
-          #rowStyle = list(background = "rgba(0, 0, 0, 0.05)"),
           theme = reactable::reactableTheme(borderWidth = '1px',
             borderColor = '#00000040'),
           selection = "multiple",
@@ -642,7 +715,6 @@ validateTableServer <- function(id, sfObject, success, con) {
           defaultPageSize = 5,
           bordered = TRUE,
           resizable = TRUE,
-          #striped = TRUE,
           compact = TRUE
         )
       })
@@ -670,7 +742,7 @@ validateTableServer <- function(id, sfObject, success, con) {
 #'
 #' @export
 #'
-#' @name validateTable
+#' @name validatePairs
 #'
 validatePairsUi <- function(id) {
   ns <- shiny::NS(id)
@@ -680,25 +752,23 @@ validatePairsUi <- function(id) {
 }
 #' @export
 #'
-#' @rdname validateTable
+#' @rdname validatePairs
 validatePairsServer <- function(id, sfObject, success, con) {
   shiny::moduleServer(
     id,
     function(input, output, session) {
-      validationResults <- shiny::reactive({
-        shiny::req(sfObject())
-      })
       output$pairValidation <- reactable::renderReactable({
         shiny::req(sfObject())
         success$depend()
         geometryType <- guess_geometry_type(sfObject())
         stagingTable <- staging_tables()[[geometryType]]
-        sfObject <- tryCatch(
-          filter_in_db(con = con,
-            tableName = stagingTable, sfObject = sfObject()),
-          error = identity)
-        transectIds <- c(sort(unique(stats::na.omit(sfObject$transect_id))))
+        transectsTable <- transects_tables()[[geometryType]]
+        dbTransectIds <- select_distinct_transect_ids(con = con,
+          stagingTable = stagingTable, transectsTable = transectsTable)
+        transectIds <- c(sort(unique(stats::na.omit(sfObject()[['transect_id']])
+          )))
         transectIds <- transectIds[transectIds != 'incidental']
+        transectIds <- unique(c(transectIds, dbTransectIds))
         validatePairs <- vapply(
           validate_sample_pairs(transectIds = transectIds), FUN = getElement,
           FUN.VALUE = character(1), name = 'messages')
@@ -735,7 +805,6 @@ validatePairsServer <- function(id, sfObject, success, con) {
         pageSizeOptions = c(5, 10, 25, 50),
         defaultPageSize = 5,
         bordered = TRUE,
-        #striped = TRUE,
         resizable = TRUE,
         compact = TRUE
         )