+++ inst/PEMcollectr/global.R
@@ -1,3 +1,4 @@
 library(shiny)
 library(PEMcollectr)
+options(shiny.maxRequestSize = 60 * 1024 ^ 2)
 con <- connect_pg('PEM')